Split Computer Handler to multiple files
