+++ coverage.out
@@ -1,17 +1,12 @@
 mode: set
-greenbone-case-study/docs/docs.go:401.13,403.2 1 0
+greenbone-case-study/main.go:16.13,27.2 6 0
 greenbone-case-study/database/db.go:14.18,19.16 2 0
 greenbone-case-study/database/db.go:19.16,21.3 1 0
 greenbone-case-study/database/db.go:23.2,24.15 2 0
 greenbone-case-study/database/db.go:27.35,32.16 2 0
 greenbone-case-study/database/db.go:32.16,34.3 1 0
 greenbone-case-study/database/db.go:35.2,36.17 2 0
-greenbone-case-study/main.go:16.13,27.2 6 0
-greenbone-case-study/routes/route.go:11.34,12.40 1 0
-greenbone-case-study/routes/route.go:12.40,14.3 1 0
-greenbone-case-study/routes/route.go:16.2,16.46 1 0
-greenbone-case-study/routes/route.go:16.46,18.3 1 0
-greenbone-case-study/routes/route.go:19.2,26.68 8 0
+greenbone-case-study/docs/docs.go:401.13,403.2 1 0
 greenbone-case-study/models/computer.go:25.36,26.69 1 0
 greenbone-case-study/models/computer.go:26.69,28.3 1 0
 greenbone-case-study/models/computer.go:29.2,29.13 1 0
@@ -34,56 +29,61 @@ greenbone-case-study/models/computer.go:91.16,93.3 1 0
 greenbone-case-study/models/computer.go:95.2,97.38 2 0
 greenbone-case-study/models/computer.go:97.38,99.3 1 0
 greenbone-case-study/models/computer.go:100.2,100.12 1 0
+greenbone-case-study/routes/route.go:12.45,13.40 1 0
+greenbone-case-study/routes/route.go:13.40,15.3 1 0
+greenbone-case-study/routes/route.go:17.2,17.46 1 0
+greenbone-case-study/routes/route.go:17.46,19.3 1 0
+greenbone-case-study/routes/route.go:21.2,30.12 9 0
 greenbone-case-study/database/db.go:14.18,19.16 2 0
 greenbone-case-study/database/db.go:19.16,21.3 1 0
 greenbone-case-study/database/db.go:23.2,24.15 2 0
 greenbone-case-study/database/db.go:27.35,32.16 2 1
 greenbone-case-study/database/db.go:32.16,34.3 1 0
 greenbone-case-study/database/db.go:35.2,36.17 2 1
-greenbone-case-study/handlers/computer.go:22.41,25.48 2 1
-greenbone-case-study/handlers/computer.go:25.48,29.3 1 1
-greenbone-case-study/handlers/computer.go:31.2,31.26 1 1
-greenbone-case-study/handlers/computer.go:31.26,35.3 1 1
-greenbone-case-study/handlers/computer.go:37.2,38.16 2 1
-greenbone-case-study/handlers/computer.go:38.16,39.50 1 1
-greenbone-case-study/handlers/computer.go:39.50,42.4 1 1
-greenbone-case-study/handlers/computer.go:44.3,44.17 1 1
-greenbone-case-study/handlers/computer.go:44.17,48.4 1 1
-greenbone-case-study/handlers/computer.go:51.2,51.53 1 1
-greenbone-case-study/handlers/computer.go:64.42,69.16 4 1
-greenbone-case-study/handlers/computer.go:69.16,70.40 1 1
-greenbone-case-study/handlers/computer.go:70.40,74.4 1 1
-greenbone-case-study/handlers/computer.go:77.2,77.48 1 1
-greenbone-case-study/handlers/computer.go:89.42,94.16 3 1
-greenbone-case-study/handlers/computer.go:94.16,98.3 1 1
-greenbone-case-study/handlers/computer.go:100.2,100.49 1 1
-greenbone-case-study/handlers/computer.go:114.45,119.16 4 1
-greenbone-case-study/handlers/computer.go:119.16,120.40 1 1
-greenbone-case-study/handlers/computer.go:120.40,124.4 1 1
-greenbone-case-study/handlers/computer.go:127.2,128.16 2 1
-greenbone-case-study/handlers/computer.go:128.16,132.3 1 1
-greenbone-case-study/handlers/computer.go:134.2,134.55 1 1
-greenbone-case-study/handlers/computer.go:150.45,154.63 3 1
-greenbone-case-study/handlers/computer.go:154.63,156.3 1 1
-greenbone-case-study/handlers/computer.go:158.2,159.45 2 1
-greenbone-case-study/handlers/computer.go:159.45,161.3 1 1
-greenbone-case-study/handlers/computer.go:163.2,171.16 7 1
-greenbone-case-study/handlers/computer.go:171.16,173.33 1 1
-greenbone-case-study/handlers/computer.go:173.33,178.4 1 1
-greenbone-case-study/handlers/computer.go:180.2,180.25 1 1
-greenbone-case-study/handlers/computer.go:200.41,207.63 5 1
-greenbone-case-study/handlers/computer.go:207.63,209.3 1 1
-greenbone-case-study/handlers/computer.go:211.2,213.58 2 1
-greenbone-case-study/handlers/computer.go:213.58,215.3 1 0
-greenbone-case-study/handlers/computer.go:216.2,216.25 1 1
-greenbone-case-study/handlers/computer.go:232.47,239.16 5 1
-greenbone-case-study/handlers/computer.go:239.16,243.3 1 1
-greenbone-case-study/handlers/computer.go:245.2,245.49 1 1
-greenbone-case-study/handlers/computer.go:263.43,268.63 3 1
-greenbone-case-study/handlers/computer.go:268.63,270.3 1 1
-greenbone-case-study/handlers/computer.go:272.2,274.58 2 1
-greenbone-case-study/handlers/computer.go:274.58,276.3 1 0
-greenbone-case-study/handlers/computer.go:277.2,277.25 1 1
+greenbone-case-study/handlers/computers/assign.go:28.33,35.63 5 1
+greenbone-case-study/handlers/computers/assign.go:35.63,37.3 1 1
+greenbone-case-study/handlers/computers/assign.go:39.2,41.58 2 1
+greenbone-case-study/handlers/computers/assign.go:41.58,43.3 1 0
+greenbone-case-study/handlers/computers/assign.go:44.2,44.25 1 1
+greenbone-case-study/handlers/computers/by_employee.go:24.47,31.16 5 1
+greenbone-case-study/handlers/computers/by_employee.go:31.16,35.3 1 1
+greenbone-case-study/handlers/computers/by_employee.go:37.2,37.49 1 1
+greenbone-case-study/handlers/computers/create.go:21.33,24.48 2 1
+greenbone-case-study/handlers/computers/create.go:24.48,28.3 1 1
+greenbone-case-study/handlers/computers/create.go:30.2,30.26 1 1
+greenbone-case-study/handlers/computers/create.go:30.26,34.3 1 1
+greenbone-case-study/handlers/computers/create.go:36.2,37.16 2 1
+greenbone-case-study/handlers/computers/create.go:37.16,38.50 1 1
+greenbone-case-study/handlers/computers/create.go:38.50,41.4 1 1
+greenbone-case-study/handlers/computers/create.go:43.3,43.17 1 1
+greenbone-case-study/handlers/computers/create.go:43.17,47.4 1 1
+greenbone-case-study/handlers/computers/create.go:50.2,50.53 1 1
+greenbone-case-study/handlers/computers/delete.go:21.37,26.16 4 1
+greenbone-case-study/handlers/computers/delete.go:26.16,27.40 1 1
+greenbone-case-study/handlers/computers/delete.go:27.40,31.4 1 1
+greenbone-case-study/handlers/computers/delete.go:34.2,35.16 2 1
+greenbone-case-study/handlers/computers/delete.go:35.16,39.3 1 1
+greenbone-case-study/handlers/computers/delete.go:41.2,41.55 1 1
+greenbone-case-study/handlers/computers/index.go:19.33,24.16 3 1
+greenbone-case-study/handlers/computers/index.go:24.16,28.3 1 1
+greenbone-case-study/handlers/computers/index.go:30.2,30.49 1 1
+greenbone-case-study/handlers/computers/show.go:20.34,25.16 4 1
+greenbone-case-study/handlers/computers/show.go:25.16,26.40 1 1
+greenbone-case-study/handlers/computers/show.go:26.40,30.4 1 1
+greenbone-case-study/handlers/computers/show.go:33.2,33.48 1 1
+greenbone-case-study/handlers/computers/unassign.go:25.35,30.63 3 1
+greenbone-case-study/handlers/computers/unassign.go:30.63,32.3 1 1
+greenbone-case-study/handlers/computers/unassign.go:34.2,36.58 2 1
+greenbone-case-study/handlers/computers/unassign.go:36.58,38.3 1 0
+greenbone-case-study/handlers/computers/unassign.go:39.2,39.25 1 1
+greenbone-case-study/handlers/computers/update.go:24.37,28.63 3 1
+greenbone-case-study/handlers/computers/update.go:28.63,30.3 1 1
+greenbone-case-study/handlers/computers/update.go:32.2,33.45 2 1
+greenbone-case-study/handlers/computers/update.go:33.45,35.3 1 1
+greenbone-case-study/handlers/computers/update.go:37.2,45.16 7 1
+greenbone-case-study/handlers/computers/update.go:45.16,47.33 1 1
+greenbone-case-study/handlers/computers/update.go:47.33,52.4 1 1
+greenbone-case-study/handlers/computers/update.go:54.2,54.25 1 1
 greenbone-case-study/models/computer.go:25.36,26.69 1 1
 greenbone-case-study/models/computer.go:26.69,28.3 1 1
 greenbone-case-study/models/computer.go:29.2,29.13 1 1
